Fix DataType issue in SQL models

+++ back_end_SQL/migrations/20220906132347-create-quote.js
@@ -20,16 +20,16 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       open: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       high: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       low: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       close: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       volume: {
         type: Sequelize.INTEGER
